Half hearted attempt to convert to real project. Fix for counting syllables in words containing erhua that is actually pronounced (ex 婴儿).
