+++ .gitignore
@@ -1,0 +1,6 @@
+build/
+dist/
+node_modules/
+*.bak
+*~
+.editorconfig
