+++ src/index.ts
@@ -1,0 +1,3 @@
+import {initUserScript} from './duchinese-helper.userscript.js';
+
+initUserScript();
